backend/requirements.txt
# --- Framework ---
fastapi
uvicorn[standard]
python-dotenv
pydantic
python-multipart

# --- Configuración ---
pydantic-settings

# --- Seguridad ---
pyjwt
passlib[bcrypt]
bcrypt==4.0.1
slowapi  # Rate limiting
python-jose[cryptography]  # JWT tokens

# --- Base de datos y Cache ---
pymysql
cryptography  # Requerido por PyMySQL
sqlalchemy
alembic
redis
celery
# --- Async Utilities ---
nest_asyncio

# --- IA y LLM ---
openai>=1.54.0  # Cliente requerido para OpenAI
tenacity>=8.0.0  # Para retry logic en llamadas a LLM
copilotkit  # SDK oficial para integración con CopilotKit

# --- GCP Services ---
google-cloud-secret-manager
google-cloud-storage
google-cloud-tasks
google-cloud-logging
google-cloud-aiplatform==1.40.0
google-cloud-documentai==2.20.1
google-cloud-language==2.13.0
google-generativeai==0.8.3

# --- HTTP Client (para servicio RAG externo) ---
httpx

# --- Procesamiento de Documentos ---
pandas
openpyxl
python-docx
python-pptx
pypdf
pdfplumber
reportlab  # Para generar PDFs con formato profesional
